enviot/requirements.txt
ansimarkup==1.4.0
appdirs==1.4.3
astroid==2.1.0
attrs==18.2.0
better-exceptions-fork==0.2.1.post6
black==18.9b0
Click==7.0
colorama==0.4.1
Flask==1.0.2
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
loguru==0.2.3
MarkupSafe==1.1.0
mccabe==0.6.1
paho-mqtt==1.4.0
pluginbase==0.7
Pygments==2.3.1
pylint==2.2.2
pymongo==3.7.2
pytz==2018.7
six==1.12.0
toml==0.10.0
typed-ast==1.1.1
tzlocal==1.5.1
Werkzeug==0.14.1
wrapt==1.10.11
